refactor WIP 2
