+++ src/features/select-pokemon-type/domain/ports/PokemonTypesRepository.ts
@@ -1,0 +1,5 @@
+import { PokemonTypeItem } from "../entities/PokemonTypeItem";
+
+export interface PokemonTypesRepository {
+  findAll(): Promise<PokemonTypeItem[]>;
+}
